[HUDI-7275] Separate use of HoodieTimelineTimeZone.UTC and LOCAL in tests to prevent infinite loops
